+++ app/feature/admin/components/create-link-dialog.tsx
@@ -1,52 +1,48 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { Button } from '~/components/ui/button';
 import { Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle } from '~/components/ui/dialog';
 import { Input } from '~/components/ui/input';
-import { DialogProps, Link } from '../types';
+import { DialogProps } from './types';
 
-interface EditLinkDialogProps extends DialogProps {
-  link: Link | null;
-  onSubmit: (values: { id: number; title: string; url: string }) => void;
+interface CreateLinkDialogProps extends DialogProps {
+  defaultValues?: {
+    title: string;
+    url: string;
+  };
+  onSubmit: (values: { title: string; url: string }) => void;
 }
 
-export function EditLinkDialog({ isOpen, onOpenChange, link, onSubmit }: EditLinkDialogProps) {
-  const [values, setValues] = useState(link || { id: 0, title: '', url: '' });
-
-  useEffect(() => {
-    if (link) {
-      setValues(link);
-    }
-  }, [link]);
+export function CreateLinkDialog({ isOpen, onOpenChange, defaultValues, onSubmit }: CreateLinkDialogProps) {
+  const [values, setValues] = useState(defaultValues || { title: '', url: '' });
 
   const handleSubmit = () => {
-    if (link) {
-      onSubmit(values);
-    }
+    onSubmit(values);
+    setValues({ title: '', url: '' });
   };
 
   return (
     <Dialog open={isOpen} onOpenChange={onOpenChange}>
       <DialogContent>
         <DialogHeader>
-          <DialogTitle>Edit Link</DialogTitle>
+          <DialogTitle>Add New Link</DialogTitle>
         </DialogHeader>
         <div className='space-y-4 py-4'>
           <div className='space-y-2'>
-            <label htmlFor='edit-title' className='text-sm font-medium text-gray-700 dark:text-gray-300'>
+            <label htmlFor='title' className='text-sm font-medium text-gray-700 dark:text-gray-300'>
               Title
             </label>
             <Input
-              id='edit-title'
+              id='title'
               value={values.title}
               onChange={(e) => setValues((prev) => ({ ...prev, title: e.target.value }))}
             />
           </div>
           <div className='space-y-2'>
-            <label htmlFor='edit-url' className='text-sm font-medium text-gray-700 dark:text-gray-300'>
+            <label htmlFor='url' className='text-sm font-medium text-gray-700 dark:text-gray-300'>
               URL
             </label>
             <Input
-              id='edit-url'
+              id='url'
               type='url'
               value={values.url}
               onChange={(e) => setValues((prev) => ({ ...prev, url: e.target.value }))}
@@ -54,10 +50,16 @@ export function EditLinkDialog({ isOpen, onOpenChange, link, onSubmit }: EditLin
           </div>
         </div>
         <DialogFooter>
-          <Button variant='outline' onClick={() => onOpenChange(false)}>
+          <Button
+            variant='outline'
+            onClick={() => {
+              onOpenChange(false);
+              setValues({ title: '', url: '' });
+            }}
+          >
             Cancel
           </Button>
-          <Button onClick={handleSubmit}>Update</Button>
+          <Button onClick={handleSubmit}>Create</Button>
         </DialogFooter>
       </DialogContent>
     </Dialog>
