+++ server/index.ts
@@ -12,18 +12,11 @@ const app = new Hono<{ Bindings: Bindings }>().basePath('/api');
 app.use(csrf());
 
 const route = app
-  .get(
-    '/app-links',
-    // cache({
-    //   cacheName: 'app-links',
-    //   cacheControl: 'max-age=10800',
-    // }),
-    async (c) => {
-      const { results } = await c.env.DB.prepare('SELECT * FROM app_links').all<AppLink>();
+  .get('/app-links', async (c) => {
+    const { results } = await c.env.DB.prepare('SELECT * FROM app_links').all<AppLink>();
 
-      return c.json(results);
-    },
-  )
+    return c.json(results);
+  })
   .post('/app-links/create', async (c) => {
     const { name, url } = await c.req.json<{ name: string; url: string }>();
